feat: Line을 조회할수없는경우 예외 처리 구현

+++ src/main/java/subway/domain/line/service/LineService.java
@@ -18,11 +18,19 @@ public class LineService {
     }
 
     public List<LineEntity> findAllLine() {
-        return lineDao.findAll();
+        Optional<List<LineEntity>> findLines = lineDao.findAll();
+        if (findLines.isEmpty()) {
+            throw new IllegalArgumentException("노선이 존재하지 않습니다.");
+        }
+        return findLines.get();
     }
 
     public LineEntity findLineById(Long id) {
-        return lineDao.findById(id);
+        Optional<LineEntity> findLine = lineDao.findById(id);
+        if (findLine.isEmpty()) {
+            throw new IllegalArgumentException("해당 ID의 노선이 존재하지 않습니다.");
+        }
+        return findLine.get();
     }
 
     public LineEntity saveLine(final LineRequest request) {
@@ -34,10 +42,18 @@ public class LineService {
     }
 
     public void updateLine(final Long id, final LineRequest lineUpdateRequest) {
+        Optional<LineEntity> findLine = lineDao.findById(id);
+        if (findLine.isEmpty()) {
+            throw new IllegalArgumentException("해당 ID의 노선이 존재하지 않습니다.");
+        }
         lineDao.update(new LineEntity(id, lineUpdateRequest.getName(), lineUpdateRequest.getColor()));
     }
 
     public void deleteLineById(final Long id) {
+        Optional<LineEntity> findLine = lineDao.findById(id);
+        if (findLine.isEmpty()) {
+            throw new IllegalArgumentException("해당 ID의 노선이 존재하지 않습니다.");
+        }
         lineDao.deleteById(id);
     }
 }
